+++ R/sentomeasures_main.R
@@ -105,7 +105,7 @@ ctr_agg <- function(howWithin = "proportional", howDocs = "equal_weight", howTim
     err <- c(err, paste0(howDocs, " is no current option for aggregation across documents."))
   }
   if (lag == 1) {
-    warning("The argument choice 'lag = 1' implies no time aggregation. We have kept a dummy weighting scheme 'dummyTime'.")
+    message("The choice 'lag = 1' implies no time aggregation, so we added a dummy weighting scheme 'dummyTime'.")
     howTime <- "own"
     weights <- data.frame(dummyTime = 1)
   }
@@ -257,8 +257,9 @@ sento_measures <- function(sento_corpus, lexicons, ctr) {
 #' sentiment measures. Can also be used to aggregate sentence-level sentiment scores into
 #' document-level sentiment scores. This function is called within the \code{\link{sento_measures}} function.
 #'
-#' @param x a \code{sentiment} object created using \code{\link{compute_sentiment}} (from a
-#' \code{sento_corpus} object), or an output from \code{\link{as.sentiment}}.
+#' @param x a \code{sentiment} object created using \code{\link{compute_sentiment}} (computed from a
+#' \code{sento_corpus} object, or any sentence-level \code{sentiment} object given \code{do.full = FALSE}), or an
+#' output from \code{\link{as.sentiment}}.
 #' @param ctr output from a \code{\link{ctr_agg}} call. The \code{howWithin} and \code{nCore} elements are ignored.
 #' @param do.full if \code{do.full = TRUE} (by default), does entire aggregation up to a \code{sento_measures}
 #' object, else only goes from sentence-level to document-level. Ignored if no \code{"sentence_id"} column in
@@ -307,10 +308,18 @@ aggregate.sentiment <- function(x, ctr, do.full = TRUE, ...) {
   weightingParamTime <- ctr$time$weightingParam
   by <- weightingParamTime$by
 
-  if ("sentence_id" %in% colnames(x)) {
+  cols <- colnames(x)
+  if ("sentence_id" %in% cols) {
     x <- aggregate_sentences(x, how = howDocs, weightingParamDocs = weightingParamDocs)
     if (do.full == FALSE) return(x)
+    if (do.full == TRUE && !("date" %in% colnames(x))) {
+      warning("Aggregation only performed until document-level, since no 'date' column present.")
+      return(x)
+    }
+  } else {
+    if (!("date" %in% cols)) stop("A document-level sentiment input should have a 'date' column for full aggregation.")
   }
+
   aggDocs <- aggregate_docs(x, by = by, how = howDocs, weightingParamDocs = weightingParamDocs)
   aggDocs$ctr <- ctr
   sento_measures <- aggregate_time(aggDocs, how = howTime, weightingParamTime = weightingParamTime)
@@ -320,14 +329,18 @@ aggregate.sentiment <- function(x, ctr, do.full = TRUE, ...) {
 
 aggregate_sentences <- function(sentiment, how, weightingParamDocs) {
   wc <- sentiment[, .(word_count)]
-  dates <- sentiment[, .(date)]
   do.ignoreZeros <- weightingParamDocs$do.ignoreZeros
   alphaExpDocs <- weightingParamDocs$alphaExpDocs
+
   weights <- weights_across(sentiment, how, do.ignoreZeros, alphaExpDocs, by = "id")
-  sw <- data.table::data.table(id = sentiment[["id"]], sentiment[, -1:-4] * weights, wc, dates)
-  s <- sw[, lapply(.SD, function(x)
-    sum(x, na.rm = TRUE)), by = c("id", "date")] # implicitly assumes all id and date combinations are unique
-  data.table::setcolorder(s, c("id", "date", "word_count"))
+
+  if ("date" %in% colnames(sentiment)) {
+    sw <- data.table::data.table(id = sentiment[["id"]], sentiment[, .(date)], wc, sentiment[, -1:-4] * weights)
+    s <- sw[, lapply(.SD, sum, na.rm = TRUE), by = c("id", "date")] # assumes all id and date combinations are unique
+  } else {
+    sw <- data.table::data.table(id = sentiment[["id"]], wc, sentiment[, -1:-3] * weights)
+    s <- sw[, lapply(.SD, function(x) sum(x, na.rm = TRUE)), by = "id"]
+  }
   class(s) <- c("sentiment", class(s))
   s
 }
@@ -494,8 +507,11 @@ peakdates <- function(sento_measures, n = 10, type = "both", do.average = FALSE)
 
 weights_across <- function(s, how = "proportional", do.ignoreZeros = TRUE, alpha = 0.1, by = "date") {
 
-  if ("id" %in% colnames(s) && !"id" %in% by) s <- s[, !"id"]
-  if ("sentence_id" %in% colnames(s) && !"sentence_id" %in% by) s <- s[, !"sentence_id"][, !"date"]
+  if ("id" %in% colnames(s) && !("id" %in% by)) s <- s[, !"id"]
+  if ("sentence_id" %in% colnames(s) && !("sentence_id" %in% by)) {
+    if ("date" %in% colnames(s)) s <- s[, !"sentence_id"][, !"date"]
+    else s <- s[, !"sentence_id"]
+  }
 
   if (how == "equal_weight") {
     if (do.ignoreZeros == TRUE) {
@@ -524,7 +540,7 @@ weights_across <- function(s, how = "proportional", do.ignoreZeros = TRUE, alpha
       weights <- s[, w := (1 / word_count) / sum(1 / word_count, na.rm = TRUE), by = eval(by)][, "w"]
       weights <- weights[, colnames(s)[-c(1:2)] := weights][, -1]
     }
-  }  else if (how == "exponential") {
+  } else if (how == "exponential") {
     # exponential w.r.t. words in document vs. total words in all documents per date
     if (do.ignoreZeros == TRUE) {
       docsIn <- s[, lapply(.SD, function(x)
@@ -548,7 +564,7 @@ weights_across <- function(s, how = "proportional", do.ignoreZeros = TRUE, alpha
                    by = eval(by)][, "w"]
       weights <- weights[, colnames(s)[-c(1:2)] := weights][, -1]
     }
-  }
+  } else stop("Weighting scheme not recognized.")
 
   weights
 }